finish the project1a

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,7 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        int quotient = (float)top / bottom;
         int result = Math.round(quotient);
         return result;
     }
